signal-worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
numpy==1.26.4
pillow==10.4.0
scipy==1.14.1
PyWavelets==1.7.0
httpx==0.27.2
pydantic==2.9.0
